# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "handlers.CodeSummarizeRequest": {
        "properties": {
          "workspace": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.CodeSummarizeResponse": {
        "properties": {
          "errors": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "processed": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "handlers.SummarizeRequest": {
        "properties": {
          "force": {
            "type": "boolean"
          },
          "limit": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "workspace": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.SummarizeResponse": {
        "properties": {
          "errors": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "summarized": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "handlers.backlinkItem": {
        "properties": {
          "collection": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.backlinksResponse": {
        "properties": {
          "doc_id": {
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/handlers.backlinkItem"
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "handlers.impactNode": {
        "properties": {
          "depth": {
            "type": "integer"
          },
          "edge_type": {
            "type": "string"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.impactRequest": {
        "properties": {
          "edge_type": {
            "type": "string"
          },
          "max_depth": {
            "type": "integer"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.impactResponse": {
        "properties": {
          "impacted": {
            "items": {
              "$ref": "#/components/schemas/handlers.impactNode"
            },
            "type": "array"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "handlers.resolveResponse": {
        "properties": {
          "query": {
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/handlers.resolveResult"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "handlers.resolveResult": {
        "properties": {
          "id": {
            "type": "string"
          },
          "match": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFToken": {
        "description": "Required on write endpoints; enforced by middleware.CSRF.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "WorkspaceAuth": {
        "description": "Workspace hash via query param or JSON body; resolved by workspaceMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      },
      "WorkspaceRegisteredAuth": {
        "description": "Workspace must be already-registered (write-path gate); enforced by workspaceRegisteredMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Self-describing OpenAPI 3.0 spec for nano-brain's non-MCP REST API",
    "title": "nano-brain REST API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/code/summarize": {
      "post": {
        "description": "Runs one batch of code-symbol summarization for the workspace, subject to daily budget limits",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.CodeSummarizeRequest"
              }
            }
          },
          "description": "Code summarize options",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.CodeSummarizeResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Summarize pending code symbols for a workspace",
        "tags": [
          "code-summarize"
        ]
      }
    },
...
    "/api/v1/graph/impact": {
      "post": {
        "description": "Returns nodes that would be impacted (transitively depend on) a given node, up to max_depth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.impactRequest"
              }
            }
          },
          "description": "Impact query",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.impactResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "Impact analysis for a graph node",
        "tags": [
          "graph"
        ]
      }
    },
...
    "/api/v1/links/resolve": {
      "get": {
        "description": "Resolves a query string as either a document UUID or a title match",
        "parameters": [
          {
            "description": "Document ID or title to resolve",
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.resolveResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "Resolve a link query to document IDs",
        "tags": [
          "links"
        ]
      }
    },
    "/api/v1/links/{doc_id}/backlinks": {
      "get": {
        "description": "Returns documents that link to the given doc_id, paginated",
        "parameters": [
          {
            "description": "Target document ID",
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max results (default 20, max 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Result offset",
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.backlinksResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "List documents that link to a document",
        "tags": [
          "links"
        ]
      }
    },
...
    "/api/v1/summarize": {
      "post": {
        "description": "Summarizes up to `limit` session documents for a workspace via the configured summarizer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.SummarizeRequest"
              }
            }
          },
          "description": "Summarize options",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.SummarizeResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Summarize pending session documents for a workspace",
        "tags": [
          "summarize"
        ]
      }
    },
...
  }
}